% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normdrawandfill.R
\name{normdrawandfill}
\alias{normdrawandfill}
\title{Load a Matrix}
\usage{
normdrawandfill(mean = 0, sd = 1, points = 200, linecolor = "orange",
  fillcolor = "blue", fillafter = NULL, fillbefore = NULL,
  fillbetween = NULL, fillwithin = NULL, thickness = 40, xlabel = NULL,
  ylabel = NULL, axiscol = "red")
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
A function to draw Normal distribution density curve (bell curve) and highlight different areas of distribution
by filling the area with chosen color and displaying Probability percentages
Created by Sulekha Aloorravi
}
